Create PersonLinkField
